pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-register"
version = "0.1.0"
description = "Automated registration with Qwen token acquisition for OpenClaw"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.40.0",
    "httpx>=0.25.0",
    "customtkinter>=5.2.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
auto-register = "auto_register.main:main"
